refactor(test): move server tests to root __tests__ directory
Moved server integration tests from src/server/__tests__ to root __tests__
directory for better organization and cleaner separation of test files from
source code. This follows a common convention where integration tests are
kept at the project root while component tests remain alongside components.

Changes:
- Moved ExpressServer.test.js to __tests__/
- Moved API route tests to __tests__/routes/api/
- Updated vitest.config.mjs to include __tests__ directory in test patterns
- Updated import paths in test files to match new location
- Added __tests__/ to coverage exclusion list
